refactor: extract filter slider into own component
